fix: ENG-037/038/039/040/041 — clamp-before-cast, narrow exceptions, clipping warnings
- ENG-037: int8_basic/int4_basic clamp abs_max before fp16 cast (prevent NaN)
- ENG-038: generate_loop warn on allow_missing_calib=True fallback
- ENG-039: patch_model narrow _get_rope_cos_sin exception handling
- ENG-040: int8_cache add clipping ratio warning (>5% threshold)
- ENG-041: kivi_style_cache add clipping ratio warning for decode K re-quant

+++ src/quant/int4_basic.py
@@ -128,9 +128,13 @@ def quantize_symmetric_int4(
     else:
         abs_max = torch.amax(abs_reshaped, dim=-1, keepdim=True)
 
-    # Avoid division by zero
-    abs_max = abs_max.to(tensor.dtype)
+    # ENG-037: Clamp abs_max BEFORE casting to tensor.dtype to prevent
+    # fp16 precision loss from producing exact 0.0 (fp16 has limited
+    # subnormal range). If abs_max is very small and cast to fp16 first,
+    # it can become 0.0, then clamp(min=1e-5) is too late — the value is
+    # already zero and scale=0 causes NaN from round(x/0).
     abs_max = abs_max.clamp(min=1e-5)
+    abs_max = abs_max.to(tensor.dtype)
 
     # INT4 symmetric range is [-7, 7]; asymmetric uses [-8, 7] (see asymmetric_quant.py).
     scale = abs_max / 7.0
